requirements.txt
streamlit
yt-dlp
openai-whisper
langchain
langchain-google-genai
python-dotenv
